ble_cli_dongle/ble_app_interactive/requirements.txt
pyserial >= 3.4
pyte >= 0.8.0